fix: 切换图床后保存按钮不可用
