+++ src-tauri/src/manager/api/mod.rs
@@ -1,15 +1,27 @@
+pub(crate) mod delete;
+pub(crate) mod list;
+pub(crate) mod upload;
+
 use std::{path::Path, str::FromStr};
 
 use async_trait::async_trait;
 use reqwest::{
     header::{HeaderMap, HeaderName, ACCEPT},
-    Method, Response, StatusCode,
+    Method, Response,
 };
 use serde::{Deserialize, Serialize};
 use serde_json::{Map, Value};
 use tauri::Window;
 
-use crate::{error::HeaderError, http::multipart::FileKind, Error, Result};
+use crate::{error::HeaderError, http::multipart::FileKind, Result};
+
+pub(crate) use self::delete::Delete;
+pub(crate) use self::delete::{DeleteKeyKind, DeleteMethod, DeleteResponseController};
+pub(crate) use self::list::{List, ListRequestMethod, ListResponseController};
+pub(crate) use self::upload::{
+    Upload, UploadContentType, UploadResponseController, UploadResponseErrorController,
+    UploadResponseStatus, UploadResponseSuccuessController,
+};
 
 #[cfg(feature = "compress")]
 use super::CompressedFormat;
@@ -30,322 +42,6 @@ pub enum AuthMethod {
     Body {
         key: String,
     },
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-#[serde(tag = "type", rename_all = "UPPERCASE")]
-pub enum UploadContentType {
-    Json {
-        key: String,
-    },
-    Multipart {
-        file_part_name: String,
-        file_kind: FileKind,
-    },
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct UploadResponseController {
-    image_url_key: String,
-    /// 有的图床不提供缩略图
-    thumb_key: Option<String>,
-    deleted_id_key: String,
-}
-
-impl UploadResponseController {
-    pub fn new<S: Into<String>, O: Into<Option<String>>>(
-        image_url_key: S,
-        thumb_key: O,
-        deleted_id_key: S,
-    ) -> Self {
-        Self {
-            image_url_key: image_url_key.into(),
-            thumb_key: thumb_key.into(),
-            deleted_id_key: deleted_id_key.into(),
-        }
-    }
-
-    pub async fn parse(&self, response: Response) -> Result<ImageItem> {
-        let json: Value = response.json().await?;
-
-        debug!("响应体：{}", json);
-
-        let url = match json.get_value_by_keys(&self.image_url_key) {
-            Value::String(s) => s,
-            Value::Null => return Err(Error::Other("没有图片链接".to_owned())),
-            _ => return Err(Error::Other("类型错误".to_owned())),
-        };
-        let deleted_id = match json.get_value_by_keys(&self.deleted_id_key) {
-            Value::String(s) => s,
-            Value::Null => return Err(Error::Other("没有删除 id".to_owned())),
-            _ => return Err(Error::Other("类型错误".to_owned())),
-        };
-
-        match &self.thumb_key {
-            None => Ok(ImageItem {
-                url,
-                deleted_id,
-                thumb: None,
-            }),
-            Some(k) => {
-                let thumb = match json.get_value_by_keys(k) {
-                    Value::String(s) => Some(s),
-                    Value::Null => None,
-                    _ => return Err(Error::Other("类型错误".to_owned())),
-                };
-
-                Ok(ImageItem {
-                    url,
-                    deleted_id,
-                    thumb,
-                })
-            }
-        }
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct Upload {
-    path: String,
-    max_size: u8,
-    allowed_formats: Vec<AllowedImageFormat>,
-    #[cfg(feature = "compress")]
-    compressed_format: CompressedFormat,
-    content_type: UploadContentType,
-    /// 请求体中图片之外的其他部分
-    other_body: Option<Map<String, Value>>,
-    controller: UploadResponseController,
-    timeout: u64,
-}
-
-impl Upload {
-    pub fn new<T: Into<Option<u64>>, M: Into<Option<Map<String, Value>>>>(
-        url: &str,
-        max_size: u8,
-        allowed_formats: Vec<AllowedImageFormat>,
-        #[cfg(feature = "compress")] compressed_format: CompressedFormat,
-        content_type: UploadContentType,
-        other_body: M,
-        controller: UploadResponseController,
-        timeout: T,
-    ) -> Self {
-        let secs: Option<u64> = timeout.into();
-        let timeout = {
-            let secs = secs.unwrap_or(5);
-
-            if secs == 0 {
-                5
-            } else {
-                secs
-            }
-        };
-        Self {
-            path: url.into(),
-            max_size,
-            allowed_formats,
-            #[cfg(feature = "compress")]
-            compressed_format,
-            content_type,
-            controller,
-            other_body: other_body.into(),
-            timeout,
-        }
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-#[serde(tag = "type", rename_all = "UPPERCASE")]
-pub enum ListRequestMethod {
-    Get,
-    Post { body: Map<String, Value> },
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct List {
-    path: String,
-    controller: ListResponseController,
-    method: ListRequestMethod,
-}
-
-impl List {
-    pub fn new(url: &str, controller: ListResponseController, method: ListRequestMethod) -> Self {
-        Self {
-            path: url.to_owned(),
-            controller,
-            method,
-        }
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct ListResponseController {
-    items_key: String,
-    image_url_key: String,
-    thumb_key: Option<String>,
-    deleted_id_key: String,
-}
-
-impl ListResponseController {
-    pub fn new<S: Into<String>, O: Into<Option<String>>>(
-        items_key: S,
-        image_url_key: S,
-        deleted_id_key: S,
-        thumb_key: O,
-    ) -> Self {
-        Self {
-            items_key: items_key.into(),
-            image_url_key: image_url_key.into(),
-            deleted_id_key: deleted_id_key.into(),
-            thumb_key: thumb_key.into(),
-        }
-    }
-
-    async fn parse(&self, response: Response) -> Result<Vec<ImageItem>> {
-        let json: Value = response.json().await?;
-
-        let items = match json.get_value_by_keys(&self.items_key) {
-            Value::Array(arr) => arr,
-            _ => return Err(Error::Other("通过 items_key 无法获取列表".to_owned())),
-        };
-
-        println!("{:?}", items);
-
-        let mut images = Vec::with_capacity(items.len());
-
-        for item in items.iter() {
-            let url = match item.get_value_by_keys(&self.image_url_key) {
-                Value::String(s) => s,
-                _ => return Err(Error::Other("没有图片链接".to_owned())),
-            };
-            let deleted_id = match item.get_value_by_keys(&self.deleted_id_key) {
-                Value::String(s) => s,
-                _ => return Err(Error::Other("没有删除 id".to_owned())),
-            };
-
-            let image_item = match &self.thumb_key {
-                None => ImageItem {
-                    url,
-                    deleted_id,
-                    thumb: None,
-                },
-                Some(k) => {
-                    let thumb = match item.get_value_by_keys(k) {
-                        Value::String(s) => Some(s),
-                        _ => None,
-                    };
-
-                    debug!("thumb: key={} value={:?}", k, thumb);
-
-                    ImageItem {
-                        url,
-                        deleted_id,
-                        thumb,
-                    }
-                }
-            };
-
-            images.push(image_item);
-        }
-
-        debug!("图片列表：{:?}", images);
-
-        Ok(images)
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-#[serde(tag = "type", rename_all = "UPPERCASE")]
-pub enum DeleteKeyKind {
-    Path,
-    Query { key: String },
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-#[serde(tag = "type", rename_all = "UPPERCASE")]
-pub enum DeleteMethod {
-    Get {
-        kind: DeleteKeyKind,
-    },
-    Delete {
-        kind: DeleteKeyKind,
-    },
-    Post {
-        body: Map<String, Value>,
-        key: String,
-    },
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-#[serde(tag = "type", rename_all = "UPPERCASE")]
-pub enum DeleteResponseController {
-    /// status-code = 200 才成功，错误时返回 unkown
-    Status,
-    Json {
-        /// 成功与否的 key
-        key: String,
-        /// 成功的值应是多少
-        should_be: Value,
-        /// 失败时的消息 key，为 None 时如果出错则返回 unkown
-        message_key: Option<String>,
-    },
-}
-
-impl DeleteResponseController {
-    async fn parse(&self, response: Response) -> Result<()> {
-        match self {
-            DeleteResponseController::Status => {
-                if response.status() != StatusCode::OK {
-                    return Err(Error::Other("unkown".to_owned()));
-                }
-            }
-            DeleteResponseController::Json {
-                key,
-                should_be,
-                message_key,
-            } => {
-                let json: Value = response.json().await?;
-
-                let value = json.get_value_by_keys(&key);
-
-                if value.is_null() {
-                    return Err(Error::Other("无法获取删除状态值".to_owned()));
-                }
-
-                debug!("删除状态：{}", value);
-
-                if &value != should_be {
-                    match message_key {
-                        None => return Err(Error::Other("unkown".to_owned())),
-                        Some(k) => match json.get(k) {
-                            None => return Err(Error::Other("unkown".to_owned())),
-                            Some(msg) => {
-                                return Err(Error::Other(msg.as_str().unwrap().to_owned()))
-                            }
-                        },
-                    }
-                }
-            }
-        };
-
-        Ok(())
-    }
-}
-
-#[derive(Debug, Serialize, Deserialize, Clone)]
-pub struct Delete {
-    path: String,
-    method: DeleteMethod,
-    controller: DeleteResponseController,
-}
-
-impl Delete {
-    pub fn new(url: &str, method: DeleteMethod, controller: DeleteResponseController) -> Self {
-        Self {
-            path: url.to_owned(),
-            method,
-            controller,
-        }
-    }
 }
 
 #[derive(Debug, Serialize, Deserialize, Clone)]
