+++ src/pages/upload/index.tsx
@@ -156,8 +156,10 @@ const Upload = () => {
 
     if (compressListener) compressListener()
 
+    console.log(resp)
+
     if (resp.type === 'Error') {
-      if (resp.code === 'image_repeated') {
+      if (resp.code === 'REPEATED') {
         const url = 'https' + resp.detail.split('https')[1]
         messageApi.warning('图片重复上传：' + url)
 
@@ -260,7 +262,7 @@ const Upload = () => {
             ))}
 
             {images.length &&
-            typeof images[images.length - 1].status === 'object' ? (
+              typeof images[images.length - 1].status === 'object' ? (
               <div className="upload-add">
                 <PlusOutlined />
               </div>
